requirements.txt
Django==5.0.1
djangorestframework==3.14.0
mysqlclient==2.2.1
python-decouple==3.8
drf-yasg==1.21.7
django-cors-headers==4.3.1

# Testing
pytest==7.4.4
pytest-django==4.7.0
pytest-cov==4.1.0
factory-boy==3.3.0

# Code Quality
bandit==1.7.6
flake8==7.0.0
black==24.1.1